PP-281 create staff frontend template
I have created a new template for the staff frontend because it has a
much more simple layout that doesn't require something as complexed as
the govuk_template. Nor does need any of the govuk_template's
branding.

In the process of creating the new template I have:
• added assets for the new template. Image and CSS (I have
transferred over most of the CSS from the user frontend but removed
all styles that aren't needed. I may have more form styles than needed
currently. But that will be easy to refactor later on when the rest of
the frontend is complete.)
• create new template page
• update layout.html with new template and changes
• remove <main> from the pages and insert that into the template
• remove used files and includes left behind from govuk_template
• remove govuk_template assets from server.js

JIRA story -
https://payments-platform.atlassian.net/browse/PP-281