PP-282 Style generate token pages
Update generate token pages to match design and copy.

Reduce dependency on IDs appearing in markup. When IDs are present they
are often used as style hooks, which lead to specificity problems.

Simplify the label test by using an accessible markup pattern. Check
for a label with a corresponding `for` attribute. There’s no need for
the label to have an ID.