PP-587 bump to 5.7 series of node docker base images
The 5.6.0 (and indeed the slightly more lax versioned 5.6 series) aren't
being rebuilt and so don't include fixes for security issues,
specifically CVE-2016-0800 (DROWN), CVE-2016-0797 and friends[1]

So, attempting to bump to the 5.7 (_not_ using the 5.7.1 tag, which
points at the same image as of the time of committing as I don't mind
tracking the minor releases) to get updates.
Let's be honest, if we're going to stick with the official images, it
appears we're being forced onto the bleeding edge anyway.

I believe we had some issues with 5.7, although I can't remember the
specifics, let's see how this builds.

[1] https://security-tracker.debian.org/tracker/DSA-3500-1