PP-1021 Fix csrf race condition
There is a race condition where a client making two parallel requests
to selfservice, the csrf token stored into the session of one request
could not be visible in the session of the other request until the
first request completes.

The session middleware used in NodeJs Express automatically saves the
session to the store at the end of the HTTP response if the session
data has been altered. With csrf, as soon as we have validated that
the token hasn't been previously used, we add it to the session to
deny other requests coming through with that token again in the
future. Even though the session has been updated immediately, the
in-memory session data is only saved back in the store at the end of
the HTTP response which is too late and create the aforementioned race
condition.

The fix simply forces the session to be saved back to the store
directly after the valid token has been used.