PP-719 Add warning when account has no notification creds set
We authenticate all smartpay notifications using Basic Auth. If the account does not have
creds set up both at our end and the smartpay end then notifications will fail with catastrophic
consequences (payment will not be captured or refunded...). Therefore a warning message
informing the user of this situation is needed. This is what I have come up with; we can come
back to it if Stephen wants to revise.