Use ensureSessionHasCsrfSecret within enforceUser
Instead of the `appendCSRF()` function (removed as now redundant).

We have a theory that this may address some of the intermittent
smoke test failures. The `appendCSRF()` function did not flush the
session changes to the database. We think there could have been a race
condition here.

After submitting the login form to `POST /login` the
`postLogin` action of `login_controller` will redirect the user to `/`
which triggers the `user.loggedIn` route. This invokes the
`auth.enforce` middleware which in turn was invoking `appendCSRF()`.

If for any reason the csrf secret had not been set in the session at
that moment, this would attempt to generate a csrf secret, store it in
the session. However the database call to persist the session would not
be triggered until the process exited. The loggedIn route would also
trigger a server-side redirect to the `paths.user.otpLogIn` route (in
auth.enforce), and so that next page would be fetched possibly before
the database transaction had been committed. So it could read the
session before the csrfsecret was present.

The `otpLogIn` route also appears to call the middleware in the wrong
order `auth.enforceUser` should be before `csrf`.