Correct order of middleware on otpLogIn
The `csrf` function checks whether the csrf token cryptographically
matches the csrfSecret. It assumes that the secret is present in the
session already. If it is not present it will display an error mesage to
the user.
An attempt to access the otp login page without a session should
redirect the user to the login screen.