PP-1006 Use of an in-memory Sqlite database for testing
Unit test `user_test.js` was exercising the persistence of user objects to
the database by mocking Sequelize as much as this was possible. As during
development the limit of testing was quickly reached the approach has now
been changed to using a real database instead.  In order to keep those tests fast
and isolated, an In-memory Sqlite database has been weaved in. A fresh
database instance is spawned and schema created before each test.

`test_sequelize_config` module is used in tests as a replacement for
`sequelize_config` module to create an instance of sqlite used by
Sequelize.