Remove unique constraint on user email
we added the ability to have usernames separate from email addresses to
accomodate users whose email server does not allow email address
containing the `+` symbol. The intention was that a user might have
several accounts with the same email address but different usernames.

This unique constraint prevents that.

This PR removes the UNIQUE constraint on user email. We retain a regular
non-unique index in case we need to query users by email.