Tighten up matching of session during logout
Previously we were just doing a substring match aganst the session data
on the username/email. Although this will work in most circumstances it
leaves open some (probably rare) edge cases where a username might be a
substring of another users's username for example.

In order to make the test pass, I changed the test fixture used to
represnt the session to be more like the real session data.