PP-1407 separate user_service from user model
THe user model was becoming bloated and hard to understand

This commit separates out the user model, which is
the Sequelize model with a few convenience methods
for modifying state, from a new user service, which
allows you to find, send things to , authenticate
users. Hopefully by making this separation integrating
with new adminuser service should be much easier.