PP-1404 Public auth client
To help in moving to new base client I have created a publicauth
client, and wired up dev_tokens_controller to use it.

I have taken the opportunity to rething how clients work -
the interface of the new client is promise based, allowing
dev token controller to be cleaned up a bit.