PP-1798 Add GOVUK template to dependencies
Includes the addition of a script (used in
pay-frontend) to convert the standard GOVUK
Mustache template to the version that supports
inheritance.

This was added so we didn't need to use the
version with inheritance on the GOVUK template
repo (which only comes as a tarball so is harder
to manage).

See this commit for details of the process:

https://github.com/alphagov/govuk_prototype_kit/commit/e38275849c58f412eb3274998b417cf04b3f75e0