PP-1798 Make tests work with new template
The new implementation puts the 'Log out' and 'My
Services' links in a propositional navigation
partial (used by the GOVUK template). This means
the tests that found the navigation by looking for
an unordered list fail.

This makes the selector for them more specific so
they cna find the navigation and fixes some issues
with another test due to the new inheritance we're
using.