PP-1602: Add team member view to self service
This gets a list of users from adminusers based upon the
service id of the current logged in user.

This is then parsed to an set of lists for each possible role, with counts of those lists
and a total

We've added service fixtures for generating requests and responses to adminusers
for services.

Also there is a basic view of the team members of a service, which is a work in
progress.