PP-1606 Fix reliance on role Id's
We don't get the database role id's from adminusers and we cannot rely on it. Changed to use the role.name for mapping.
Also introduced an external Id to render and get user's role id from for submission. Purposely keeping it different from DB id's so that we don't expose them by chance.