PP-1798 Add IE variant stylesheets
The GOVUK template has <link> tags for IE6-8
variants of the main stylesheet. Until these are
removed we should provide the files so the network
requests are fulfilled.

(Without our using any of the associated Sass
helpers, they are essentially duplicates of
`application.css` and are only delivered to those
browsers.)