PP-1606 Add role="link" to buttons styled as links
We need these to be form submissions in order to
send a token with the request but to users they
are styled as links.

Because of this we need to ensure they are
identified as links to the accessibility API so
any assistive tech used conveys information that
matches the visuals.

Adding this role should achieve this.