PP-1930 Improve pagination buttons text
The existing text relies too much on being read
out as part of its context. For example, the
button for page 1 is just '1' which could mean a
number of things outside of a pagination block.

Users of assistive tech often view form controls
outside of their context, for example when they
ask it for all the buttons in the page. This
should help with that. I'm hiding it for visual
users as they always see the buttons in context.