PP-2001 Refactor HTML of API keys page
Includes the following changes.

The main content of this page are these keys so it
should be broken down into sections, identified
by headings.

All details follow the key:value format.
Description lists allow this to be included in the
semantics.

This is only supported in IE from version 11 so
removing.