PP-2051 Pin pact version to 1.0.0
Previously we had pact version in package.json as `^1.0.0`
(minimum of 1.0.0). We recently started seeing unit tests failin gon master
(see https://build.ci.pymnt.uk/job/pay-selfservice/job/master/4/consoleFull), which
appears due to new version of pact that was incompatible with our tests.

This commit pins the version of both pact libraries that we use. It also moves
the `pact` library to dev dependencies, as it is not needed in prod.

If we wish to upgrade pact version in future, we can do so explicitly.