PP-2001 Give transaction date fields labels
Identifies the date range fields as a fieldset and
gives labels to it (in the form of a legend) and
its fields.

Testing with a screenreader, I've had to make some
changes to make the form make sense without the
visuals. I've found it made sense to split the
date range into 'from' and 'to' groups which
required some CSS changes to support the new HTML.