PP-2085 Moved Grunt packages from dependencies to dev dependencies in package.json
We wanted to move Grunt packages from dependencies to dev dependencies
in order to follow best practises and also to decrease the chance to
be exposed to security vulnerabilities in the future.

In order to do that we needed to stop giving the choice to start the app
in either development or production mode - now production is the one in use.