PP-2001 Make login validation match GOVUK Elements
Make the login validation use GOVUK Elements
validation styles:

http://govuk-elements.herokuapp.com/errors/

This introduces a few big changes. I had to
make up a paragraph for the error summary and
hard-code error messages for the two fields as our
authentication controller only returns a single
message.

Also worth noting the CSS for these validation
styles has changed in the latest release of GOVUK
Elements so when we upgrade to that we'll need to
change the CSS used here.

The GOVUK Elements validation requires each
message is listed at the top and that only the
fields at fault are shown as errored.

We can notify users of blank fields individually
whereas validation coming from adminusers needs
to be generic. In that case we explain both fields
need to be valid.