PP-2993 - Migrating to Nunjucks
I have migrated from Mustache to Nunjucks because Mustache is no
longer maintained or used by the Node.js community, Nunjucks on
the other hand is widely used. It’s also used widely within GDS
(and also shares it's syntax with Jinja2 which Python projects use)

This first commit is a rebase an almalgamation of my inital spike
work to see how Nunjucks could work for Pay and what config was
needed to get things going.

But basically, I’ve removed the old templating packages and added
Nunjucks

Most of this is just mindless template conversion but I have renamed
and moved all the old 'token' templates into their own directory to
be inline with all the other settings templates I also renamed API
keys not 'dev tokens' as this was confusing as we don't talk about them
as tokens anywhere but here.