PP-2655 - Reusable client side form validation
Usually we write validation into the controllers
which ends up meaning we have a lot of duplication
which is a shame and is loooong.

So this is a generic browserified validation engine.
To use it anywhere you need to do a couple of things

- add `data-validate="true"` to the form you want to validate
- add `data-validate` to each of the `input/select/textarea`'s you want validated
- by default it will check that the value is not blank but you can add a space separated list of validations e.g. `data-validate="required currency"`

Currently I have only written validators for isCurrency
and isEmpty. You can add new ones to `field-validation-checks.js`

The validator expects them to return a string with the error or else false