Streamline Jenkins pipeline
This commit removes the requirement that end-to-end tests are run on
PR build. This change has been tried out already on connector and
has been successful. This removal is potentially  controversial - it makes
it possible to commit potentially broken (or incompatible) code to master.
However, I believe developer discipline and intuition should guide whether
running this before merge is necessary or not. For small changes (which
we should be encouraging) this seems way too much largely irrelevant
testing. And if you do break master, it would be much quicker to get out
a fix. If you feel need for Jenkins to run e2e against your branch before
merge, I have added a manually configurable option to be able to do that.