PP-2655 - PhantomJS doesnt like fancy ES6 stuff
We were using the spread operator [...] to convert a node list
into a standard array and as its ES6 it was converted on build.
in then used Array.from() which upset PhantomJS. We could update
PhantomJS but we're phasing it out so easier to not use the spread operator here.