PP-2604 Auto-detect whether we should use chamber
If we're running in ECS, wrap execution with chamber.

Use the presence or absence of $ECS_SERVICE as a test
for whether we're running within ECS or not.

This means we no longer need a dedicated npm entry point
in packages.json for running under chamber in ECS.
Leave it there for the time being so that existing ECS
deployments don't fall over.