PP-3239 migrating to transactions search API
Until now we've been searching using the search charges API (with the feature flag enabling to search for refunds using a UNION query).
We now have a new API for transactions search (using `/transactions` endpoint) which uses the refactored data structures. This PR migrates the selfservice search transactions/download CSV features to use the new endpoint.

- Updated connector_client to user the new endpoint
- removed the need for feature flag
- Updated tests to depict the changes as well as a new test to test multi select `state` query

Not sure if `state` param still needs sending to connector (Possibly not). Leaving it for now to remove in a later PR