PP-3505 - Making template more straight forward
When selfservice was younger all pages except the transactions bit
showed the side settings navigation. Now we're adding more functionality
this is sort of breaking down and causing problems with resuable
elements. Such as the flash error/success banner appearing in the wrong
places.

Summary of changes
- removed full_width_layout block
- added new side_navigation block for any side nav stuff
- removed display converter logic that controlled the display of the
settings navigation and dropped in side_navigation instead.