PP-3498 Results from 'State' dropdown on transactions filter should only relate to the selected payment state
Previously the state dropdown was mapping all the errors to failed. We
want to be able to search by each type of error. Needed to map the
errors to their own code and pass this to connector. Connector has
already been updated to handle these new codes. Will come back and
remove old mappings once they have been removed from the connector
response. This had to be done in a number of steps to preserve
backwards compatibility.

with @kakumara