PP-3498 Results from 'State' dropdown on transactions filter should only relate to the selected payment state
Removing unneeded code. Previously we returned failed for the connector
state of a transaction and need to use the error code to map to the
state display value. Because we wanted to search by these states we
need to send different states for each one we displayed. This means we
no longer need to use the error code to map to display states so
removing the code.