PP-3498 Handle new states with old display value
If the user does not have the new states feature flag enable they
should still see the old display values. As we have changed connector
to return the new states we need to map those new states to  the old
display values. Will all be deleted once we get rid of the feature flag.

with @kakumara