PP-3510 - display description linebreaks
Linebreaks were not showing as nunjucks strips them out by default, this
is good most of the time. So in this case we can use nunjucks filters to
show. If we convert the \n breaks to <br>'s and then escape we can see
them. Because escaping is dangerous we also strip out all tags so there
is nothing dodgy left to escape.