PP-3603 - Moving Merchant details to work at service level
As merchant details are set at the service level it should be shown on
the services page. But as it was it would show the active service rather
than the service the user thinks they might be editing.

I’ve added the externalServiceId to the path so it’s always clear which
service to edit.