PP-3686
Removed nock references from wrapper_arguments_test as some of them made no difference
to the the success or failure of a test.

I've also removed an assertion of a return value being added to a response object
as this does little more than test a third party library, rather than our own unit
of code which is the primary intention.