PP-3567 Tweaks to run selfservice in docker4mac
Appmetrics is a very os specific library. To be able to run
a locally built selfservice inside a docker container, we basically
can't run appmetrics, so I have added a flag to disable it.
It is admittedly a slightly confusing flag ('click here if you
don't want to not have the chance to win a house'), but means
appmetrics is enabled by deault, reducing amount of other apps
that need to be touched.

This also adds a `start-local` command, which runs the app with nodemon
which reloads app on file changes.