Add autocomplete="off" to OTP security code input elements
One-time passwords are by definition for one-time use, so there’s
no point having browser autocomplete remember them or make
suggestions based on past input.

We already had autocomplete="off" for the main sign in flow; this
adds it for registration of a new user and configuring the sign-in
method of an existing user.