PP-3813 Split connector contract tests
We want to run the contract tests pertaining to
transactions as part of the pipeline, and build out
connector support for those. We don't want to worry
about other tests. I have marked all those tests we
don't want to worry about for now as `selfservice-to-be`
following precedent for adminusers.