bau: Rename selfservice consumer
We have a selfservice-products pact in the pact broker but there are no
provider tests in products to verify the pact yet. This is currently causing
selfservice deploys to fail as the deploy job asks the pact broker "can i deploy
selfservice?". The broker sees there's no compatibility between selfservice and
products and fails the build. Renaming selfservice -> selfservice-to-be will
unblock us for the moment.

@oswaldquek