PP-3928 - Added autocomplete to service page
If a service has 8 or more services then we switch them over to a
slightly more condensed layout and show them an input which allows them
to jump to a specific service quickly.

This has been implemented in Javascript and will only be available for
users with JS. We have used the GOV.UK accessible autocomplete to do all
the heavy lifting.

We have a function that on selection hides all other services, leaving
just the one the user asked for.

Had to also add a little bit of config to the Gruntfile to load in the
styles for the autocomplete.