PP-3975 Filter gatewayAccountIds by external and internal ids
We were missing filtering gateway accounts for Direct Debit when we list services.
For Direct Debit currently we initialise `req.user.serviceRoles[].service.gatewayAccountIds` with external ids, but for Cards we initialise with internal ids.
We check the gateway accounts' external id first as not to overlap/skip between card payment and direct debit gateway accounts.

Commented out temporary failing test as agreed we need to merge that to unblock the current work.

with @SandorArpa