PP-4228 - Remember gateway account from last login
The getCurrentGatewayAccountId() function in auth_service.js looks up
the users Gateway account from the cookie. This cookie persists even
when logged out (for up to 30 days), so when you log back in you end up
on the same service.

But when you login it checks that you’re still a part of that service as
whilst you were gone you could have been removed.

This check was happening without the service middleware before it so it
couldnt find the service and so kicked you back to the first service in
the list.

So I added the missing middlewares and this made it work OK for users
that had the cookie.

For users with no cookies this meant the currentGatewayAccountId was set
to an empty object. Which the resolve_service middleware skipped over
and assumed meant you shouldn't be here and locked you out.

So I have added another if statement so that when there is no
cookie it defaults back to the first account in the users serviceRoles
object.

If however they happen to have no serviceRoles or are somehow trying to
access one they’re not part of they get the no rights error.

Phew