BAU - Excluding files that nodemon should ignore
Nodemon was reloading when there were changes in these files and nothing
will happen if it does, so it might as well not reload on them.

These are are assets that require running npm run compile (to run Grunt)
then this will change, application.js which then will trigger nodemon to
do reload.