PP-4472 Amend logic in resolve service
Adds a failing test to demonstrate the bug described in
https://payments-platform.atlassian.net/browse/PP-4472

Then adds the fix for this test - the logic is as follows:
In resolve_service middleware, if `req.params.externalServiceId` exists
(which means it is in the url), when we check it against the list of user's
services, we can be confident that if the user does not have access to
that service, then they should be 403ed.

If, on the other hand, the user has a cookie with a gatewayAccountId that
they no longer have access to, we do not want to 403 them at this point -
it may be because they have an old cookie, and subsequent middleware takes
care of this case.

So this commit adds the 403 in the first case.

It also returns a 403 if there is no way at all to resolve the user's service.
This previously returned a 500.