PP-4472 - Simplify gatewayAccountType function
This function exists to work out whether or not a gateway is purely card, purely direct debit or a mixture, it took mem a whilel to get my head around.

I think ideally it could do with renaming to something like onlyContainsDirectDebitGateways and onlyContainsCardGateways.

But that would require more work a refactoring more files and I am probably already beyond the scope of this ticket