PP-4472 - Add more test coverage for resolve service middleware
Tests for each of the checks that resolve services does to get a user where they need to be.

The service resolver tries to resolve via the param (for times when the user is not within a service (aka team members page) or via the cookie set with the users details.

This commit includes tests for when services can be resolved fine and for when they can’t and we must assume the user isnt authorised to be where they are.