BAU: Add request as a peer dependency
Our dependency `requestretry` moved the `request` library from a dependency to
a peer dependency in release
[2.0.0](https://github.com/FGRibreau/node-request-retry/commit/aafafe3).

A recent [PR](https://github.com/alphagov/pay-selfservice/pull/1050) bumped our
version to the latest, including this change. This adds `request` to our
dependencies to ensure the peer dependency is fulfilled and correctly installed.