PP-4448 Add request to go live link to dashboard
Replace the "Next steps to go live" link with the "Request a live
account" link for card accounts.

There are three variants of the link which are displayed under the
following circumstances:
1. If the request to go live process has not been started, the "Request
a live account" link is shown, which links to the request to go live
status page.
2. If the request to go live process has been partially completed, the
"Setting up your live account" link is shown, which links to the request
to go live status page.
3. If the request to go live process has been completed, but the account
has not yet been set up, the "Your live account" text is shown, which is
not a link.

For direct debit accounts, the "Next steps to go live" link to the
documentation is still shown, as before.