PP-4448 Modify logic for displaying links on dashboard
Move logic for deciding which links will be displayed from the Nunjucks
template to the controller. This allows us to know how many links will
be displayed, and so format them correctly into either 2 or 3 columns
with the correct borders on the links. This is required as the logic
gets a bit more complicated when we add in the "request to go live" link

Move the decision for whether to display a bottom border on the links
from the template, where we were conditionally adding a css class to
the css, using the nth-child selector.

Add Cypress tests for displaying the links.