PP-4463 Remove responsible person client-side JavaScript validation
Our current client-side (in-browser) validation framework isn’t
flexible enough to support the level of date of birth validation we
want on the Stripe responsible person page. If we had this date of
birth validation only in the server-side code then the page would
behave strangely if there were multiple errors (some would be
caught client-side, others wouldn’t be caught until the form is
actually POSTed to the server).

Remove the client-side validation for the responsible person page
to avoid the issue. The validation was only applied when the user
pressed the submit button (as opposed to when the focus leaves the
field, for example), so the server-side validation offers the same
experience, just a little slower.