PP-4904 - Add husky and lint-staged
Not everyone has linter plugins installed in their chosen editors, this means PRs fail on Jenkins for silly things like curly bracket spacing or identation.

This PR adds Husky a git hooks tool and lint-staged which make it easy to automatically lint all JS changes that have been staged for commit. If it spots errors it can fix it does or throws an error detailing what needs changing.

This will help improve the quality of code we’re shipping.