PP-4939 Fix error navigating to Organisation Details when name is set
When organisation details are incomplete, i.e. name or mandatory
address fields are not present, take the user to the edit details page
with existing details pre-filled rather than the view details page. This
fixes the 500 error thrown when the country code was not set on the view
details page.