PP-4463 Send responsible person details to Stripe
When the user successfully fills in the responsible person form,
send the details to Stripe via the Stripe SDK (2019-02-19 API
version) and tell connector to flip the flag saying that the
details have been submitted for this gateway account.

Add tests for this part of the controller (other parts of the
controller — validation, navigation back and forth between the
form and the check answers page etc. — will be tested by Cypress
tests).