PP-4939 Render error page for unhandled errors
Renamed error_handler.js to error_logger.js as this middleware just logs
the error then passes it on.
Added middleware to the end of the stack to render the GOV.UK generic
error page when an unhandled error is caught. Includes a unique log
string to make it possible to create a search for unhandled errors.

Co-authored-by: Steven Fountain <steven.fountain@digital.cabinet-office.gov.uk>