PP-4976 Use controller errors on the edit merchant details page
The controller now returns an 'errors' object with error messages as
the values, where previously there was just a boolean to indicate there
was an error.
Use these error messages in the template rather than the existing static
text.

This modifies some of the error messages, but makes them more consistent
with the design system.