PP-4949 Disable client-side validation of company number
Disable client-side JavaScript validation of company number because
it will validate the contents of the company number field even if
the user has changed their mind and said they do not have a company
number. The server-side validation, which remains in place, is more
sophisticated and does not suffer from this deficiency.

Add a test for the above scenario.

Co-Authored-By: Yancoba Thompson <yancoba.thompson@digital.cabinet-office.gov.uk>