PP-4888 Add metadata to csv download
- User research proved that using each metadata key as a column was the most useful
approach to adding metadata to the csv download. The keys are sorted in ascending
ascii order as per the default behaviour of JS list.sort() function.
- The maximum number of metadata related columns is set to 100. 100 has been used
as this seems a sensible limit when the current maximum metadata keys per charge
is set at 10. This permits some variation in key names for a service whilst preventing
the maximum number of columns in the csv download becoming too great. This is calculated
by first sorting the unique metadata keys and taking the first 100.
- Unit and ITests have been updated.