PP-5199 Remove transaction type check for displaying psp fee
- On the transaction details page, we were checking whether the
transaction_type of the charge was a 'refund' to decide whether to
display the 'PSP fee' row. However, the connector charge endpoint does
not return this field, so this check always resolves as true.
- The cypress test for this was passing because it was using an
incorrect fixture which was not validated by a pact test, which will be
amended in a separate PR.