BAU - Move phase banner out of main content block
Our phase banner was included with `mainContent`. This is bad because if people use the skip link to jump to content they actually jump to the nav. This would get annoying if you were using the skip link between every page.

It’s also just more correct.

This also means we can remove a bit of custom styling