BAU - Back links should also be before the main element
I removed the back links from the settings pages because they seem
redundant as there is the settings link in the side nav just to the
left.

Also we used a top margin all over the place but now we have things where they are supposed to be this is redundant.