PP-5560 - Yellow should only be used for focus
In the past we used yellow to denote test accounts but this was not
really kosher, but we got away with it. Now the yellow is way more
yellow it looks pretty out there so we’re switching to grey.

This will all get redesigned entirely soon so its just temporary