PP-5560 Loosen payment link validation
We were previously not allowing special characters which were
whitelisted by NAXSI. Update the validation rule to only disallow
characters that NAXSI will block, and update the error message
displayed to the user appropriately.

Also add the field validation to the "Details" field on the payment
link information screen, as this would previously not warn if
disallowed characters were entered and instead show a generic error page.