PP-5755 Add Sentry
Send error level log messages to Sentry. Using the log level to
control when messages are sent to Sentry seems intuitive and since we now
require a common logger throughout the application it offers a
convenient point to add Sentry.

Unhandled exceptions will be caught by the Sentry error middleware added
within server.js as per the guidlines when using Sentry with Express.

When initialising Sentry a beforeSend filter is added to remove the
request element from the event if it exists. The request element contains the
data from a failed request and this can include sensitive information such
as card numbers. We may want to introduce more refined filtering to allow
non-sensitive information from a failed request to go to Sentry at a
later date.

app/utils/sentry.js and more elements of app/utils/logger.js can
probably move to pay-js-commons once we're happy with the implementation.