PP-5775 Add middleware to log requests in JSON format
The middleware uses morgan to log requests in a JSON format using a
format function from pay-java-commons which produces a log line as a
JSON string.
We then parse this as JSON and give it to our custom logger, which uses
winston to produce a structured log message which also includes all the
global keys we log for all lines.