BAU Fix editing a Welsh service name
If a service has a Welsh name set, and the user goes to edit the service
name, when they click submit, the Welsh name is removed.

This was because on the edit name page there isn't a checkbox that the
controller was expecting to indicate whether to add a Welsh name.