PP-5760 Don't log error for invalid OTP code
Adminusers does some validation on the OTP code that selfservice doesn't
perform, meaning it can return a 400 if the code is not numeric or is
blank.

In selfservice we were treating this as an internal server error, when
it is actually a user input error. Change this to treat a 400 response
as a user input error.

Also don't log user input errors, only log if there was an unexpected
error.